requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
colorama>=0.4.4
tabulate>=0.8.9
pymetasploit3

PyJWT
dnspython
